requirements.txt
djangorestframework==3.14.0
#mysqlclient==1.3.14
